+++ src/components/ShoppingCard/ShoppingCardStyled.ts
@@ -12,17 +12,6 @@ const Styled = {
     width: 28px;
     height: 28px;
   `,
-  InfoContainer: styled.div`
-    display: flex;
-    flex-direction: column;
-    gap: 23px;
-
-    width: 100%;
-  `,
-  TopSection: styled.section`
-    display: flex;
-    justify-content: space-between;
-  `,
 };
 
 export default Styled;
