+++ src/components/ShoppingCard/ShoppingInfo/ShoppingInfoStyled.ts
@@ -1,0 +1,17 @@
+import styled from "styled-components";
+
+const Styled = {
+  InfoContainer: styled.div`
+    display: flex;
+    flex-direction: column;
+    gap: 23px;
+
+    width: 100%;
+  `,
+  TopSection: styled.section`
+    display: flex;
+    justify-content: space-between;
+  `,
+};
+
+export default Styled;
